Move Form* types into common-components
