+++ ui-packages/packages/form-details/src/api/FormDetailsEnvelopeApi.ts
@@ -16,7 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { FormInfo } from '@kogito-apps/forms-list';
+import {
+  Association,
+  FormInfo
+} from '@kogito-apps/components-common/dist/types';
 
 /**
  * Envelope Api
@@ -30,32 +33,4 @@ export interface FormDetailsEnvelopeApi {
     association: Association,
     formData: FormInfo
   ): Promise<void>;
-}
-
-export interface Association {
-  origin: string;
-  envelopeServerId: string;
-}
-
-export interface FormResources {
-  scripts: {
-    [key: string]: string;
-  };
-  styles: {
-    [key: string]: string;
-  };
-}
-interface FormConfiguration {
-  schema: string;
-  resources: FormResources;
-}
-export interface Form {
-  formInfo: FormInfo;
-  source: string;
-  configuration: FormConfiguration;
-}
-
-export interface FormContent {
-  source: string;
-  configuration: FormConfiguration;
 }
