+++ ui-packages/packages/forms-list/src/api/FormsListChannelApi.ts
@@ -16,7 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { FormFilter, FormInfo } from './FormsListEnvelopeApi';
+import {
+  FormFilter,
+  FormInfo
+} from '@kogito-apps/components-common/dist/types';
 /**
  * Channel Api for Forms List
  */
