+++ ui-packages/packages/forms-list/src/api/FormsListEnvelopeApi.ts
@@ -19,30 +19,12 @@
 /**
  * Envelope Api
  */
+
+import { Association } from 'packages/components-common/dist/types';
 export interface FormsListEnvelopeApi {
   /**
    * Initializes the envelope.
    * @param association
    */
   formsList__init(association: Association): Promise<void>;
-}
-
-export interface Association {
-  origin: string;
-  envelopeServerId: string;
-}
-
-export interface FormFilter {
-  formNames: string[];
-}
-
-export enum FormType {
-  HTML = 'HTML',
-  TSX = 'TSX'
-}
-
-export interface FormInfo {
-  name: string;
-  type: FormType;
-  lastModified: Date;
 }
