+++ ui-packages/packages/forms-list/src/envelope/components/FormsListUtils/FormsListUtils.tsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import Moment from 'react-moment';
 import { DataTableColumn } from '@kogito-apps/components-common/dist/components/DataTable';
-import { FormInfo } from '../../../api/FormsListEnvelopeApi';
+import { FormInfo } from '@kogito-apps/components-common/dist/types';
 import { Label } from '@patternfly/react-core';
 
 export const getFormTypeLabel = (value) => {
