+++ ui-packages/packages/runtime-tools-dev-ui-webapp/src/channel/FormsList/FormsListGatewayApi.ts
@@ -16,7 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { FormFilter, FormInfo } from '@kogito-apps/forms-list';
+import {
+  FormFilter,
+  FormInfo
+} from '@kogito-apps/components-common/dist/types';
 import { getForms } from '../apis';
 
 /* eslint-disable @typescript-eslint/no-empty-interface */
