+++ ui-packages/packages/task-form/src/envelope/components/CustomTaskFormDisplayer/CustomTaskFormDisplayer.tsx
@@ -24,13 +24,15 @@ import {
 } from '@kogito-apps/ouia-tools/dist/utils/OuiaUtils';
 import {
   EmbeddedFormDisplayer,
-  FormDisplayerApi,
+  FormDisplayerApi
+} from '@kogito-apps/form-displayer';
+import { UserTaskInstance } from '@kogito-apps/task-console-shared';
+import {
+  CustomForm,
   FormOpenedState,
   FormSubmitResponseType,
   FormOpened
-} from '@kogito-apps/form-displayer';
-import { UserTaskInstance } from '@kogito-apps/task-console-shared';
-import { CustomForm } from '@kogito-apps/components-common/dist/types';
+} from '@kogito-apps/components-common/dist/types';
 import { generateFormData } from '../utils/TaskFormDataUtils';
 import { TaskFormDriver, User } from '../../../api';
 import { Stack, StackItem } from '@patternfly/react-core/layouts/Stack';
